updated pointDF test and added operators

+++ AlgorithmsFFLib/src/MS2Ion.h
@@ -10,6 +10,7 @@
 
 #include "Error.h"
 #include "GlobalSettings.h"
+#include "PointDF.h"
 
 
 using namespace Error;
@@ -19,7 +20,7 @@ class ALGORITHMSFFLIB_EXPORTS MS2Ion {
 public:
 
     double mz = -1.0;
-    double intensity = -1.0;
+    float intensity = -1.0;
     IRT iRT = -1.0;
     QString ionLabel;
     int rank = -1;
